ex12: using more robust type long long during calculations.

+++ ex12/ft_iterative_factorial.c
@@ -6,13 +6,13 @@
 /*   By: mnenonen <mnenonen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/15 21:29:40 by mnenonen          #+#    #+#             */
-/*   Updated: 2019/10/16 19:16:37 by mnenonen         ###   ########.fr       */
+/*   Updated: 2019/10/23 22:09:12 by mnenonen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 int		ft_iterative_factorial(int nb)
 {
-	int	ans;
+	long long ans;
 
 	ans = 1;
 	if (nb < 0)
@@ -22,6 +22,6 @@ int		ft_iterative_factorial(int nb)
 		ans = ans * nb;
 		--nb;
 	}
-	nb = ans;
+	nb = (int) ans;
 	return (nb);
 }
